Add app-state selectors
